Better testing and cleanup

+++ app-modules/ai/src/Jobs/DeleteExternalAiThread.php
@@ -23,6 +23,8 @@ class DeleteExternalAiThread implements ShouldQueue
     {
         $service = $this->aiThread->assistant->model->getService();
 
-        $service->deleteThread($this->aiThread);
+        if ($service->isThreadExisting($this->aiThread)) {
+            $service->deleteThread($this->aiThread);
+        }
     }
 }
